+++ inst/include/dust/utils.hpp
@@ -34,24 +34,30 @@ size_t stride_copy(T dest, const std::vector<U>& src, size_t at,
   return at;
 }
 
+template <typename T, typename U = T>
+inline HOSTDEVICE T align_padding(const T offset, const U align) {
+  T remainder = offset % align;
+  return remainder ? align - remainder : 0;
+}
+
 #ifdef __NVCC__
 template <typename T>
-T epsilon_nvcc();
+HOSTDEVICE T epsilon_nvcc();
 
 template <>
-inline float epsilon_nvcc() {
+inline DEVICE float epsilon_nvcc() {
   return FLT_EPSILON;
 }
 
 template <>
-inline double epsilon_nvcc() {
+inline DEVICE double epsilon_nvcc() {
   return DBL_EPSILON;
 }
 #endif
 
 template <typename T>
-T epsilon() {
-#ifdef __NVCC__
+HOSTDEVICE T epsilon() {
+#ifdef __CUDA_ARCH__
   return epsilon_nvcc<T>();
 #else
   return std::numeric_limits<T>::epsilon();